Initial commit
This plugin need foo-jscript-panel.
Add a jscript panel and paste the code in it.

Create a new playlist with name "LastFM".

Right click on the panel and choose "Properties". Then configure the
Last.FM account name and the minimal number of music to keep in the
LastFM playlist.

When switching to the playlist and removing items, it will update the playlist with
recommendations from Last.FM based on configured account.

Music items are removed as they are played and new one are added (after
they are finished).